add changes after mentors review
