+++ src/components/app/app.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, {useEffect} from "react";
 import MainScreen from "../main-screen/main-screen";
 import {
   Switch,
@@ -14,23 +14,31 @@ import PrivateRoute from "../private-route/private-route";
 import {Props} from "./app-types";
 import {AppRoute} from "../../const";
 import browserHistory from "../../browser-history";
+import {useDispatch} from "react-redux";
+import {checkAuth, fetchOffersData} from "../../store/api-actions";
 
 const App: React.FC = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(checkAuth());
+    dispatch(fetchOffersData());
+  });
+
   return (
     <BrowserRouter history={browserHistory}>
       <Switch>
         <Route exact path={AppRoute.MAIN_SCREEN}>
           <MainScreen />
         </Route>
-        <Route exact path={AppRoute.LOGIN_SCREEN}>
+        <Route path={AppRoute.LOGIN_SCREEN}>
           <LoginScreen />
         </Route>
-        <PrivateRoute exact path={AppRoute.FAVORITES_SCREEN}
+        <PrivateRoute path={AppRoute.FAVORITES_SCREEN}
           renderComponent={() => <FavoritesScreen />}
           redirect={() => <Redirect to={AppRoute.LOGIN_SCREEN} />}
         />
         <Route
-          exact
           path={AppRoute.OFFER_SCREEN}
           render={({match}: Props) => <RoomScreen cardId={Number(match.params.id)} />}
         />
