+++ src/components/map/map-types.ts
@@ -1,10 +1,10 @@
 import {OfferCard, City} from "../../types";
 
 export interface mapProps {
-  offerCity?: City;
-  cards: Array<OfferCard>;
+  offerCity?: City,
+  cards: Array<OfferCard>,
   style?: {
-    height: string;
+    height: string
   };
   isMainMap?: boolean,
   activeCardId?: number
