+++ src/components/private-route/private-route-types.ts
@@ -2,5 +2,6 @@ import {ReactElement} from "react";
 
 export interface PrivateRouteProps {
   renderComponent: () => ReactElement,
-  redirect: () => ReactElement
+  redirect: () => ReactElement,
+  path: string
 }
