+++ src/index.tsx
@@ -5,10 +5,9 @@ import {createAPI} from "./services/api";
 import {Provider} from "react-redux";
 import App from "./components/app/app";
 import {rootReducer} from "./store/root-reducer";
-import {checkAuth, fetchOffersData} from "./store/api-actions";
 import {AuthorizationStatus} from "./const";
 import {redirect} from "./middlewares/redirect";
-import {requireAuthorization} from "./store/action";
+import {requireAuthorization} from "./store/actions";
 
 const api = createAPI(
     () => store.dispatch(requireAuthorization(AuthorizationStatus.NO_AUTH))
@@ -23,9 +22,6 @@ const store = configureStore({
       },
     }).concat(redirect)
 });
-
-store.dispatch(checkAuth());
-store.dispatch(fetchOffersData());
 
 ReactDom.render(
     <Provider store={store}>
