+++ src/middlewares/redirect.ts
@@ -5,7 +5,7 @@ import {RootStateType} from "../store/root-reducer";
 
 export const redirect: Middleware<RootStateType> = (_store) => (next) => (action) => {
   if (action.type === ActionType.REDIRECT_TO_ROUTE) {
-    browserHistory.push(action.url);
+    browserHistory.push(action.payload);
   }
 
   return next(action);
